requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
branca==0.5.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
Django==3.2.15
django-crispy-forms==1.14.0
django-osm-field==0.5.0
django-upgrade==1.7.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
filelock==3.7.1
folium==0.12.1.post1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.10.2
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
lightgbm==3.3.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
np==1.0.2
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyarrow==9.0.0
pycodestyle==2.9.1
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.7
pyzmq==23.0.0
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.12.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenize-rt==4.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.1
traitlets==5.2.2.post1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
virtualenv==20.16.3
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
xgboost==1.6.1
zipp==3.8.1